Fixes from self-review
