+++ src/spec-lang/tc/internal_types/index.ts
@@ -1,0 +1,2 @@
+export * from "./builtin_struct_type";
+export * from "./function_set_type";
