+++ test/unit/interposing.spec.ts
@@ -24,7 +24,7 @@ import {
 } from "../../src/instrumenter";
 import { cook } from "../../src/rewriter";
 import { single } from "../../src/util";
-import { getTypeCtxAndTarget, toAst } from "../integration/utils";
+import { getTarget, getTypeCtxAndTarget, toAst } from "../integration/utils";
 import { getCallGraph } from "../../src/instrumenter/callgraph";
 import { getCHA } from "../../src/instrumenter/cha";
 import { InstrumentationContext } from "../../src/instrumenter/instrumentation_context";
@@ -319,7 +319,7 @@ contract Foo {
     ] of goodSamples) {
         it(`Interpose on ${contractName}.${funName} in #${fileName}`, () => {
             const [sources, reader, files, compilerVersion] = toAst(fileName, content);
-            const [, target] = getTypeCtxAndTarget([contractName, funName], sources);
+            const target = getTarget([contractName, funName], sources);
             const fun: FunctionDefinition = target as FunctionDefinition;
             const factory = new ASTNodeFactory(reader.context);
 
@@ -441,7 +441,7 @@ contract Foo is __scribble_ReentrancyUtils {
         it(`Instrument ${contractName} in #${fileName}`, () => {
             const [sources, reader, files, compilerVersion] = toAst(fileName, content);
 
-            const [, target] = getTypeCtxAndTarget([contractName, undefined], sources);
+            const target = getTarget([contractName, undefined], sources);
             const contract: ContractDefinition = target as ContractDefinition;
             const factory = new ASTNodeFactory(reader.context);
             const contractInstrumenter = new ContractInstrumenter();
